(wip) Begin refactor of tm.js
The TM class constructor was a bit confusing because it configured
`this` but then returned new on another class.

This new implementation uses a factory pattern, and also delegates the
validation of the options (TM needs type, TM needs URL, etc) to the
concrete instances.

This is also lays groundwork for hopefully removing hardcoded strings
like 'tm2' and 'tm3' which occuring in src/db/seeds,
api/tests/services, and pages/admin.